pyproject.toml
[project]
name = "own-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.71.2",
    "openai-agents[viz]>=0.0.16",
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "pywin32>=310",
    "rich>=14.0.0",
    "streamlit>=1.45.1",
]
